refactor: collapse narrativeGenerator.prompt.* sub-files into one module
The prompt enhancer was split across 8 files where the root file was just
re-exports. Each sub-module was a single small function with duplicated
imports. Merged context/lore/tone/inventory/itemAcquisition/itemLoss/types
into narrativeGenerator.prompt.ts; kept personalization separate (cohesive
~300-line module with its own test).
